Prevent google-auth "empty workspace" warning
The "google-github-actions/auth" action shows this warning is the workspace is empty:

The "create_credentials_file" option is true, but the current GitHub workspace is empty. Did you forget to use "actions/checkout" before this step? If you do not intend to share authentication with future steps in this job, set "create_credentials_file" to false.

This warning is irrelevant for us since we do not need the workspace in the "apply-terraform" job, since it downloads the Terraform plan from the previous job (post approval) to gain its Terraform code & plan.

Therefore, to prevent the warning, the artifact downloading is moved up to run before the auth step, so that the workspace is not empty, which would suppress the warning.